telegram-bot-v2/py-clob-server/requirements.txt
# 24/7 Server Requirements: FastAPI + Telegram Bot
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PostgreSQL Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Background task scheduling
apscheduler>=3.10.0

# Telegram Bot dependencies
python-telegram-bot>=20.0

# Twitter Bot dependencies
requests-oauthlib>=1.3.1

# Core dependencies (reuse from parent bot)
requests>=2.28.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
cryptography>=3.4.0

# AI Categorization
openai>=1.0.0

# For Railway deployment
gunicorn>=21.2.0

# Core blockchain dependencies from parent project
web3>=6.0.0
eth-account>=0.8.0
eth-keys>=0.4.0

# py_clob_client dependencies (required by local package)
py-order-utils>=0.3.2
poly-eip712-structs>=0.0.1
eth-utils>=4.1.1

# PostgreSQL Database Support - CRITICAL FOR PERSISTENT STORAGE
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Solana Bridge Dependencies
solana>=0.30.0
solders>=0.18.0
anchorpy>=0.18.0
base58>=2.1.0
pynacl>=1.5.0
aiohttp>=3.8.0

# Redis Cache for Performance Optimization
redis>=5.0.0

# Timezone support for leaderboard scheduler
pytz>=2023.3

# Force rebuild Tue Oct 14 21:36:46 CEST 2025
